修复 AMP 混合精度下 predictions 容器 dtype 不匹配崩溃
开 use_amp=True 时模型 forward 输出为 float16，但 predictions 容器用
torch.zeros 默认创建 float32，line 621 predictions[mask] = bond_logits_flat
因 dtype 不匹配直接 RuntimeError 崩溃。

修复：predictions 与空分支 bond_logits_flat 的 dtype 跟随 node_features，
AMP 下自动为 half，非 AMP 下为 float32，两种模式都兼容。

触发场景：default.yaml device.use_amp=True 时跑任意消融都会崩。
与消融逻辑无关，是 forward 路径原有的 dtype bug。

+++ graph_transform/models/graph_transformer.py
@@ -585,7 +585,9 @@ class GraphTransformer(nn.Module):
             max_bonds = int(precomputed_max_bonds)
         else:
             max_bonds = int(bond_counts.max().item()) if batch_size > 0 else 0
-        predictions = torch.zeros(batch_size, max_bonds, device=node_features.device)
+        # dtype 跟随 node_features：AMP 下模型输出为 half，predictions 容器也需 half，
+        # 否则 predictions[valid_bond_mask] = bond_logits_flat 会因 dtype 不匹配而崩。
+        predictions = torch.zeros(batch_size, max_bonds, device=node_features.device, dtype=node_features.dtype)
 
         if max_bonds > 0:
             bond_positions = torch.arange(max_bonds, device=node_features.device)
@@ -620,7 +622,7 @@ class GraphTransformer(nn.Module):
             bond_logits_flat = self.bond_head(bond_features).squeeze(-1)
             predictions[valid_bond_mask] = bond_logits_flat
         else:
-            bond_logits_flat = torch.empty(0, device=node_features.device)
+            bond_logits_flat = torch.empty(0, device=node_features.device, dtype=node_features.dtype)
 
         if timing_enabled:
             _maybe_sync(node_features.device, True)
